chore(web): extract Notifications panel to /theguard/connections/notifications
Phase 2 Chunk A panel 1 of 4. Moves NotificationsCard + Slack setup
checklist out of the Guard settings tabs into a dedicated route +
features/guard/connections/notifications/Panel.tsx.

Settings page shim redirects legacy ?tab=notifications bookmarks to the
new URL. No API, schema, or UX behavior changes — copy-move only.
